+++ src/misc/common.cpp
@@ -192,11 +192,10 @@ const std::string& getBinaryLocation() {
 
 /**
  */
-int Split(std::vector<std::string>& vecteur, std::string chaine, char separateur) {
+size_t Split(std::vector<std::string>& vecteur, std::string chaine, char separateur) {
   vecteur.clear();
   std::string::size_type stTemp = chaine.find(separateur);
-  while(stTemp != std::string::npos)
-  {
+  while(stTemp != std::string::npos) {
     vecteur.push_back(chaine.substr(0, stTemp));
     chaine = chaine.substr(stTemp + 1);
     stTemp = chaine.find(separateur);
